Initial commit
A tested version of the script and workflow that converts exports from
Paribas BNP to the YNAB format.

Note that the "Payee" field is derived from the "Memo" field by parsing
the relevant information our based on matching one of several patterns.
This is a heuristic based on some sample data that I received.